+++ src/Commands/SiteDesigns/RemoveSiteDesign.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.Online.SharePoint.TenantAdministration;
 using Microsoft.SharePoint.Client;
-
 using PnP.PowerShell.Commands.Base;
 using PnP.PowerShell.Commands.Base.PipeBinds;
 using System.Management.Automation;
@@ -18,9 +17,15 @@ namespace PnP.PowerShell.Commands
 
         protected override void ExecuteCmdlet()
         {
+            var siteDesign = Identity.GetTenantSiteDesign(Tenant);
+            if(siteDesign == null)
+            {
+                throw new PSArgumentException("Site design provided through the Identity parameter could not be found", nameof(Identity));
+            }
+
             if (Force || ShouldContinue(Properties.Resources.RemoveSiteDesign, Properties.Resources.Confirm))
             {
-                Tenant.DeleteSiteDesign(Identity.Id);
+                Tenant.DeleteSiteDesign(siteDesign.Id);
                 ClientContext.ExecuteQueryRetry();
             }
         }